Cloud_Computing_Master_Note_Encrypter_Final/requirements.txt
flask
pylint
black
mysql.connector 
pycryptodome
rsa
gunicorn